\name{stopping}

\alias{stopping}
\docType{data}
\title{Stopping}

\description{
This data set concerns the stopping distances for cars travelling at different speeds. There are replicate values for some speeds. The objective is to find a model for these data.
}

\usage{data("stopping")}
\format{
  A data frame with 63 observations on the following 2 variables.
  \describe{
    \item{\code{Speed}}{travelling speed (miles per hour).}
    \item{\code{Distance}}{stopping distance (feet).}
    }
}


\source{
Snee, R.D. (1986) An alternative approach to fitting models when re-expression of the
response is useful. \emph{Journal of Quality Technology}, 18,211-225.

A Handbook of Small Data Sets (p.354) by Hand, Daly, McConway, Lunn, Ostrowski. Chapman & Hall.
}